[FlexNG] Don't let min size go negative
We were subtracting border_padding when it wasn't present if the item
had min-size:auto but didn't qualify for the special flex min-size:auto
behavior because, for instance, it has overflow: hidden.

This fixes the telemetry_perf_unittests failure FlexNG was hitting on
the bots, which was caused by a recorded google docs page.

Change-Id: Ia6bde490651daf63e4ca77f9525870987bb9fe5c