Align <pattern> and <*Gradient> attribute collection
Currently we stop collection of the pattern attributes if one of the
<pattern>s that is inherited from is not in the layout tree (but still
use the result). For gradients this cause the generation of the paint
server to fail (triggering fallback if available).

So consider a missing LayoutObject the same as a missing element in the
various Collect...Attributes implementations, essentially ignoring the
reference.
This also aligns our behavior with that of Gecko.

Add const qualification to the CollectGradientAttributes() methods.

Bug: 109212

Change-Id: Ie8a22664c4dc37ebc10fb02f18e1176a9d41a0e0