[PerformanceObserver] Fix observe() edge cases
Currently, two calls to observe with the same 'type' will result in a
console error, whereas the spec implies that this is valid, with the
later call overriding the initial one. This CL fixes that. It also fixes
a problem where disconnect would not clear the filter_options_, as shown
in https://github.com/w3c/performance-timeline/pull/161. This would not
be a problem with 'entryTypes' because this variable would be replaced
on every call to observe(). Tests are added for these two cases.

Bug: 1052509
Change-Id: I75865b2925e2fb1a7e7fd64d7e979907c3b7650c